chore: move contract class struct to snapi
