+++ crates/blockifier/src/execution/contract_class.rs
@@ -1,35 +1,23 @@
 use std::collections::{HashMap, HashSet};
-use std::ops::Deref;
 use std::sync::Arc;
 
-use cairo_lang_casm;
-use cairo_lang_casm::hints::Hint;
-use cairo_lang_starknet_classes::casm_contract_class::{CasmContractClass, CasmContractEntryPoint};
+use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use cairo_lang_starknet_classes::NestedIntList;
-use cairo_vm::serde::deserialize_program::{
-    ApTracking,
-    FlowTrackingData,
-    HintParams,
-    ReferenceManager,
-};
 use cairo_vm::types::builtin_name::BuiltinName;
 use cairo_vm::types::errors::program_errors::ProgramError;
-use cairo_vm::types::program::Program;
-use cairo_vm::types::relocatable::MaybeRelocatable;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use itertools::Itertools;
-use serde::de::Error as DeserializationError;
-use serde::{Deserialize, Deserializer};
-use starknet_api::core::EntryPointSelector;
-use starknet_api::deprecated_contract_class::{
-    sn_api_to_cairo_vm_program,
-    ContractClass as DeprecatedContractClass,
-    EntryPoint,
-    EntryPointOffset,
-    EntryPointType,
-    Program as DeprecatedProgram,
+use starknet_api::contract_class::{
+    ClassInfo,
+    ContractClass,
+    ContractClassV0,
+    ContractClassV0Inner,
+    ContractClassV1,
+    ContractClassV1Inner,
+    EntryPointV1,
 };
-use starknet_types_core::felt::Felt;
+use starknet_api::core::EntryPointSelector;
+use starknet_api::deprecated_contract_class::EntryPointType;
 
 use super::execution_utils::poseidon_hash_many_cost;
 use crate::abi::abi_utils::selector_from_name;
@@ -50,39 +38,32 @@ pub mod test;
 
 pub type ContractClassResult<T> = Result<T, ContractClassError>;
 
-#[derive(Clone, Debug, Eq, PartialEq, derive_more::From)]
-pub enum ContractClass {
-    V0(ContractClassV0),
-    V1(ContractClassV1),
+pub trait ContractClassExt {
+    fn constructor_selector(&self) -> Option<EntryPointSelector>;
+    fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources;
+    fn get_visited_segments(
+        &self,
+        visited_pcs: &HashSet<usize>,
+    ) -> Result<Vec<usize>, TransactionExecutionError>;
+    fn bytecode_length(&self) -> usize;
 }
 
-impl TryFrom<starknet_api::contract_class::ContractClass> for ContractClass {
-    type Error = ProgramError;
-
-    fn try_from(
-        contract_class: starknet_api::contract_class::ContractClass,
-    ) -> Result<Self, Self::Error> {
-        let starknet_api::contract_class::ContractClass::V1(contract_class_v1) = contract_class;
-        Ok(ContractClass::V1(contract_class_v1.try_into()?))
-    }
-}
-
-impl ContractClass {
-    pub fn constructor_selector(&self) -> Option<EntryPointSelector> {
+impl ContractClassExt for ContractClass {
+    fn constructor_selector(&self) -> Option<EntryPointSelector> {
         match self {
             ContractClass::V0(class) => class.constructor_selector(),
             ContractClass::V1(class) => class.constructor_selector(),
         }
     }
 
-    pub fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources {
+    fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources {
         match self {
             ContractClass::V0(class) => class.estimate_casm_hash_computation_resources(),
             ContractClass::V1(class) => class.estimate_casm_hash_computation_resources(),
         }
     }
 
-    pub fn get_visited_segments(
+    fn get_visited_segments(
         &self,
         visited_pcs: &HashSet<usize>,
     ) -> Result<Vec<usize>, TransactionExecutionError> {
@@ -94,7 +75,7 @@ impl ContractClass {
         }
     }
 
-    pub fn bytecode_length(&self) -> usize {
+    fn bytecode_length(&self) -> usize {
         match self {
             ContractClass::V0(class) => class.bytecode_length(),
             ContractClass::V1(class) => class.bytecode_length(),
@@ -103,17 +84,13 @@ impl ContractClass {
 }
 
 // V0.
-#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq)]
-pub struct ContractClassV0(pub Arc<ContractClassV0Inner>);
-impl Deref for ContractClassV0 {
-    type Target = ContractClassV0Inner;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
+trait ContractClassV0Ext {
+    fn constructor_selector(&self) -> Option<EntryPointSelector>;
+    fn n_entry_points(&self) -> usize;
+    fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources;
 }
 
-impl ContractClassV0 {
+impl ContractClassV0Ext for ContractClassV0 {
     fn constructor_selector(&self) -> Option<EntryPointSelector> {
         Some(self.entry_points_by_type[&EntryPointType::Constructor].first()?.selector)
     }
@@ -136,51 +113,48 @@ impl ContractClassV0 {
             builtin_instance_counter: HashMap::from([(BuiltinName::pedersen, hashed_data_size)]),
         }
     }
+}
 
-    pub fn n_builtins(&self) -> usize {
+pub trait ContractClassV0PubExt {
+    fn n_builtins(&self) -> usize;
+    fn bytecode_length(&self) -> usize;
+    fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV0, ProgramError>;
+}
+
+impl ContractClassV0PubExt for ContractClassV0 {
+    fn n_builtins(&self) -> usize {
         self.program.builtins_len()
     }
 
-    pub fn bytecode_length(&self) -> usize {
+    fn bytecode_length(&self) -> usize {
         self.program.data_len()
     }
 
-    pub fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV0, ProgramError> {
+    fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV0, ProgramError> {
         let contract_class: ContractClassV0Inner = serde_json::from_str(raw_contract_class)?;
         Ok(ContractClassV0(Arc::new(contract_class)))
     }
 }
 
-#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq)]
-pub struct ContractClassV0Inner {
-    #[serde(deserialize_with = "deserialize_program")]
-    pub program: Program,
-    pub entry_points_by_type: HashMap<EntryPointType, Vec<EntryPoint>>,
-}
-
-impl TryFrom<DeprecatedContractClass> for ContractClassV0 {
-    type Error = ProgramError;
-
-    fn try_from(class: DeprecatedContractClass) -> Result<Self, Self::Error> {
-        Ok(Self(Arc::new(ContractClassV0Inner {
-            program: sn_api_to_cairo_vm_program(class.program)?,
-            entry_points_by_type: class.entry_points_by_type,
-        })))
-    }
-}
-
 // V1.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ContractClassV1(pub Arc<ContractClassV1Inner>);
-impl Deref for ContractClassV1 {
-    type Target = ContractClassV1Inner;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
+trait ContractClassV1Ext {
+    fn constructor_selector(&self) -> Option<EntryPointSelector>;
+    fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources;
+    fn get_visited_segments(
+        &self,
+        visited_pcs: &HashSet<usize>,
+    ) -> Result<Vec<usize>, TransactionExecutionError>;
 }
 
-impl ContractClassV1 {
+pub trait ContractClassV1PubExt {
+    fn bytecode_length(&self) -> usize;
+    fn get_entry_point(&self, call: &CallEntryPoint) -> Result<EntryPointV1, PreExecutionError>;
+    fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV1, ProgramError>;
+    #[cfg(any(feature = "testing", test))]
+    fn empty_for_testing() -> Self;
+}
+
+impl ContractClassV1Ext for ContractClassV1 {
     fn constructor_selector(&self) -> Option<EntryPointSelector> {
         Some(self.0.entry_points_by_type[&EntryPointType::Constructor].first()?.selector)
     }
@@ -189,7 +163,7 @@ impl ContractClassV1 {
     /// This is an empiric measurement of several bytecode lengths, which constitutes as the
     /// dominant factor in it.
     fn estimate_casm_hash_computation_resources(&self) -> ExecutionResources {
-        estimate_casm_hash_computation_resources(&self.bytecode_segment_lengths)
+        estimate_casm_hash_computation_resources(self.bytecode_segment_lengths())
     }
 
     // Returns the set of segments that were visited according to the given visited PCs.
@@ -199,21 +173,15 @@ impl ContractClassV1 {
         visited_pcs: &HashSet<usize>,
     ) -> Result<Vec<usize>, TransactionExecutionError> {
         let mut reversed_visited_pcs: Vec<_> = visited_pcs.iter().cloned().sorted().rev().collect();
-        get_visited_segments(&self.bytecode_segment_lengths, &mut reversed_visited_pcs, &mut 0)
+        get_visited_segments(self.bytecode_segment_lengths(), &mut reversed_visited_pcs, &mut 0)
     }
-
-    pub fn bytecode_length(&self) -> usize {
+}
+impl ContractClassV1PubExt for ContractClassV1 {
+    fn bytecode_length(&self) -> usize {
         self.program.data_len()
     }
 
-    pub fn bytecode_segment_lengths(&self) -> &NestedIntList {
-        &self.bytecode_segment_lengths
-    }
-
-    pub fn get_entry_point(
-        &self,
-        call: &CallEntryPoint,
-    ) -> Result<EntryPointV1, PreExecutionError> {
+    fn get_entry_point(&self, call: &CallEntryPoint) -> Result<EntryPointV1, PreExecutionError> {
         if call.entry_point_type == EntryPointType::Constructor
             && call.entry_point_selector != selector_from_name(CONSTRUCTOR_ENTRY_POINT_NAME)
         {
@@ -236,22 +204,22 @@ impl ContractClassV1 {
         }
     }
 
-    pub fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV1, ProgramError> {
+    fn try_from_json_string(raw_contract_class: &str) -> Result<ContractClassV1, ProgramError> {
         let casm_contract_class: CasmContractClass = serde_json::from_str(raw_contract_class)?;
-        let contract_class: ContractClassV1 = casm_contract_class.try_into()?;
+        let contract_class = casm_contract_class.try_into()?;
 
         Ok(contract_class)
     }
 
     /// Returns an empty contract class for testing purposes.
     #[cfg(any(feature = "testing", test))]
-    pub fn empty_for_testing() -> Self {
-        Self(Arc::new(ContractClassV1Inner {
-            program: Default::default(),
-            entry_points_by_type: Default::default(),
-            hints: Default::default(),
-            bytecode_segment_lengths: NestedIntList::Leaf(0),
-        }))
+    fn empty_for_testing() -> Self {
+        Self(Arc::new(ContractClassV1Inner::new_for_testing(
+            Default::default(),
+            Default::default(),
+            Default::default(),
+            NestedIntList::Leaf(0),
+        )))
     }
 }
 
@@ -346,198 +314,44 @@ fn get_visited_segments(
     Ok(res)
 }
 
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ContractClassV1Inner {
-    pub program: Program,
-    pub entry_points_by_type: HashMap<EntryPointType, Vec<EntryPointV1>>,
-    pub hints: HashMap<String, Hint>,
-    bytecode_segment_lengths: NestedIntList,
+pub trait ClassInfoExt: Sized {
+    fn bytecode_length(&self) -> usize;
+    fn contract_class(&self) -> ContractClass;
+    fn sierra_program_length(&self) -> usize;
+    fn abi_length(&self) -> usize;
+    fn code_size(&self) -> usize;
+    fn new(
+        contract_class: &ContractClass,
+        sierra_program_length: usize,
+        abi_length: usize,
+    ) -> ContractClassResult<Self>;
 }
 
-#[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
-pub struct EntryPointV1 {
-    pub selector: EntryPointSelector,
-    pub offset: EntryPointOffset,
-    pub builtins: Vec<BuiltinName>,
-}
-
-impl EntryPointV1 {
-    pub fn pc(&self) -> usize {
-        self.offset.0
-    }
-}
-
-impl TryFrom<starknet_api::contract_class::ContractClassV1> for ContractClassV1 {
-    type Error = ProgramError;
-
-    fn try_from(
-        contract_class: starknet_api::contract_class::ContractClassV1,
-    ) -> Result<Self, Self::Error> {
-        let starknet_api::contract_class::ContractClassV1::Casm(casm_contract_class) =
-            contract_class;
-        casm_contract_class.try_into()
-    }
-}
-
-impl TryFrom<CasmContractClass> for ContractClassV1 {
-    type Error = ProgramError;
-
-    fn try_from(class: CasmContractClass) -> Result<Self, Self::Error> {
-        let data: Vec<MaybeRelocatable> = class
-            .bytecode
-            .into_iter()
-            .map(|x| MaybeRelocatable::from(Felt::from(x.value)))
-            .collect();
-
-        let mut hints: HashMap<usize, Vec<HintParams>> = HashMap::new();
-        for (i, hint_list) in class.hints.iter() {
-            let hint_params: Result<Vec<HintParams>, ProgramError> =
-                hint_list.iter().map(hint_to_hint_params).collect();
-            hints.insert(*i, hint_params?);
-        }
-
-        // Collect a sting to hint map so that the hint processor can fetch the correct [Hint]
-        // for each instruction.
-        let mut string_to_hint: HashMap<String, Hint> = HashMap::new();
-        for (_, hint_list) in class.hints.iter() {
-            for hint in hint_list.iter() {
-                string_to_hint.insert(serde_json::to_string(hint)?, hint.clone());
-            }
-        }
-
-        let builtins = vec![]; // The builtins are initialize later.
-        let main = Some(0);
-        let reference_manager = ReferenceManager { references: Vec::new() };
-        let identifiers = HashMap::new();
-        let error_message_attributes = vec![];
-        let instruction_locations = None;
-
-        let program = Program::new(
-            builtins,
-            data,
-            main,
-            hints,
-            reference_manager,
-            identifiers,
-            error_message_attributes,
-            instruction_locations,
-        )?;
-
-        let mut entry_points_by_type = HashMap::new();
-        entry_points_by_type.insert(
-            EntryPointType::Constructor,
-            convert_entry_points_v1(class.entry_points_by_type.constructor),
-        );
-        entry_points_by_type.insert(
-            EntryPointType::External,
-            convert_entry_points_v1(class.entry_points_by_type.external),
-        );
-        entry_points_by_type.insert(
-            EntryPointType::L1Handler,
-            convert_entry_points_v1(class.entry_points_by_type.l1_handler),
-        );
-
-        let bytecode_segment_lengths = class
-            .bytecode_segment_lengths
-            .unwrap_or_else(|| NestedIntList::Leaf(program.data_len()));
-
-        Ok(Self(Arc::new(ContractClassV1Inner {
-            program,
-            entry_points_by_type,
-            hints: string_to_hint,
-            bytecode_segment_lengths,
-        })))
-    }
-}
-
-// V0 utilities.
-
-/// Converts the program type from SN API into a Cairo VM-compatible type.
-pub fn deserialize_program<'de, D: Deserializer<'de>>(
-    deserializer: D,
-) -> Result<Program, D::Error> {
-    let deprecated_program = DeprecatedProgram::deserialize(deserializer)?;
-    sn_api_to_cairo_vm_program(deprecated_program)
-        .map_err(|err| DeserializationError::custom(err.to_string()))
-}
-
-// V1 utilities.
-
-// TODO(spapini): Share with cairo-lang-runner.
-fn hint_to_hint_params(hint: &cairo_lang_casm::hints::Hint) -> Result<HintParams, ProgramError> {
-    Ok(HintParams {
-        code: serde_json::to_string(hint)?,
-        accessible_scopes: vec![],
-        flow_tracking_data: FlowTrackingData {
-            ap_tracking: ApTracking::new(),
-            reference_ids: HashMap::new(),
-        },
-    })
-}
-
-fn convert_entry_points_v1(external: Vec<CasmContractEntryPoint>) -> Vec<EntryPointV1> {
-    external
-        .into_iter()
-        .map(|ep| EntryPointV1 {
-            selector: EntryPointSelector(Felt::from(ep.selector)),
-            offset: EntryPointOffset(ep.offset),
-            builtins: ep
-                .builtins
-                .into_iter()
-                .map(|builtin| BuiltinName::from_str(&builtin).expect("Unrecognized builtin."))
-                .collect(),
-        })
-        .collect()
-}
-
-#[derive(Clone, Debug)]
-// TODO(Ayelet,10/02/2024): Change to bytes.
-pub struct ClassInfo {
-    contract_class: ContractClass,
-    sierra_program_length: usize,
-    abi_length: usize,
-}
-
-impl TryFrom<starknet_api::contract_class::ClassInfo> for ClassInfo {
-    type Error = ProgramError;
-
-    fn try_from(class_info: starknet_api::contract_class::ClassInfo) -> Result<Self, Self::Error> {
-        let starknet_api::contract_class::ClassInfo {
-            contract_class,
-            sierra_program_length,
-            abi_length,
-        } = class_info;
-
-        let contract_class: ContractClass = contract_class.try_into()?;
-        Ok(Self { contract_class, sierra_program_length, abi_length })
-    }
-}
-
-impl ClassInfo {
-    pub fn bytecode_length(&self) -> usize {
+impl ClassInfoExt for ClassInfo {
+    fn bytecode_length(&self) -> usize {
         self.contract_class.bytecode_length()
     }
 
-    pub fn contract_class(&self) -> ContractClass {
+    fn contract_class(&self) -> ContractClass {
         self.contract_class.clone()
     }
 
-    pub fn sierra_program_length(&self) -> usize {
+    fn sierra_program_length(&self) -> usize {
         self.sierra_program_length
     }
 
-    pub fn abi_length(&self) -> usize {
+    fn abi_length(&self) -> usize {
         self.abi_length
     }
 
-    pub fn code_size(&self) -> usize {
+    fn code_size(&self) -> usize {
         (self.bytecode_length() + self.sierra_program_length())
             // We assume each felt is a word.
             * eth_gas_constants::WORD_WIDTH
             + self.abi_length()
     }
 
-    pub fn new(
+    fn new(
         contract_class: &ContractClass,
         sierra_program_length: usize,
         abi_length: usize,
