+++ crates/blockifier/src/execution/entry_point_execution.rs
@@ -10,11 +10,12 @@ use cairo_vm::vm::runners::builtin_runner::BuiltinRunner;
 use cairo_vm::vm::runners::cairo_runner::{CairoArg, CairoRunner, ExecutionResources};
 use cairo_vm::vm::security::verify_secure_runner;
 use num_traits::{ToPrimitive, Zero};
+use starknet_api::contract_class::{ContractClassV1, EntryPointV1};
 use starknet_api::felt;
 use starknet_types_core::felt::Felt;
 
 use crate::execution::call_info::{CallExecution, CallInfo, Retdata};
-use crate::execution::contract_class::{ContractClassV1, EntryPointV1};
+use crate::execution::contract_class::ContractClassV1PubExt;
 use crate::execution::entry_point::{
     CallEntryPoint,
     EntryPointExecutionContext,
