+++ crates/blockifier/src/fee/actual_cost_test.rs
@@ -4,6 +4,7 @@ use starknet_types_core::felt::Felt;
 
 use crate::context::BlockContext;
 use crate::execution::call_info::{CallExecution, CallInfo, MessageToL1, OrderedL2ToL1Message};
+use crate::execution::contract_class::ClassInfoExt;
 use crate::fee::eth_gas_constants;
 use crate::fee::gas_usage::{
     get_consumed_message_to_l2_emissions_cost,
