+++ crates/blockifier/src/state/state_api.rs
@@ -1,5 +1,6 @@
 use std::collections::{HashMap, HashSet};
 
+use starknet_api::contract_class::ContractClass;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::state::StorageKey;
 use starknet_types_core::felt::Felt;
@@ -7,7 +8,6 @@ use starknet_types_core::felt::Felt;
 use super::cached_state::{ContractClassMapping, StateMaps};
 use crate::abi::abi_utils::get_fee_token_var_address;
 use crate::abi::sierra_types::next_storage_key;
-use crate::execution::contract_class::ContractClass;
 use crate::state::errors::StateError;
 
 pub type StateResult<T> = Result<T, StateError>;
