+++ crates/blockifier/src/test_utils/struct_impls.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use serde_json::Value;
 use starknet_api::block::{BlockNumber, BlockTimestamp};
+use starknet_api::contract_class::{ContractClassV0, ContractClassV1};
 use starknet_api::core::{ChainId, ContractAddress, Nonce, PatriciaKey};
 use starknet_api::transaction::{Calldata, Fee, TransactionHash, TransactionVersion};
 use starknet_api::{calldata, contract_address, felt, patricia_key};
@@ -14,7 +15,7 @@ use crate::blockifier::block::{BlockInfo, GasPrices};
 use crate::bouncer::{BouncerConfig, BouncerWeights};
 use crate::context::{BlockContext, ChainInfo, FeeTokenAddresses, TransactionContext};
 use crate::execution::call_info::{CallExecution, CallInfo, Retdata};
-use crate::execution::contract_class::{ContractClassV0, ContractClassV1};
+use crate::execution::contract_class::{ContractClassV0PubExt, ContractClassV1PubExt};
 use crate::execution::entry_point::{
     CallEntryPoint,
     EntryPointExecutionContext,
@@ -220,15 +221,19 @@ impl CallExecution {
 
 // Contract loaders.
 
-impl ContractClassV0 {
-    pub fn from_file(contract_path: &str) -> Self {
+pub trait FromFileExt {
+    fn from_file(contract_path: &str) -> Self;
+}
+
+impl FromFileExt for ContractClassV0 {
+    fn from_file(contract_path: &str) -> Self {
         let raw_contract_class = get_raw_contract_class(contract_path);
         Self::try_from_json_string(&raw_contract_class).unwrap()
     }
 }
 
-impl ContractClassV1 {
-    pub fn from_file(contract_path: &str) -> Self {
+impl FromFileExt for ContractClassV1 {
+    fn from_file(contract_path: &str) -> Self {
         let raw_contract_class = get_raw_contract_class(contract_path);
         Self::try_from_json_string(&raw_contract_class).unwrap()
     }
