+++ crates/native_blockifier/src/py_transaction.rs
@@ -1,16 +1,16 @@
 use std::collections::BTreeMap;
 
 use blockifier::execution::contract_class::{
-    ClassInfo,
-    ContractClass,
-    ContractClassV0,
-    ContractClassV1,
+    ClassInfoExt,
+    ContractClassV0PubExt,
+    ContractClassV1PubExt,
 };
 use blockifier::transaction::account_transaction::AccountTransaction;
 use blockifier::transaction::transaction_execution::Transaction;
 use blockifier::transaction::transaction_types::TransactionType;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
+use starknet_api::contract_class::{ClassInfo, ContractClass, ContractClassV0, ContractClassV1};
 use starknet_api::transaction::{Resource, ResourceBounds};
 use starknet_api::StarknetApiError;
