+++ crates/starknet_api/src/contract_class.rs
@@ -1,20 +1,236 @@
-use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
+use std::collections::HashMap;
+use std::ops::Deref;
+use std::sync::Arc;
 
-/// Compiled contract class.
-#[derive(Clone, Debug, Eq, PartialEq)]
+use cairo_lang_casm::hints::Hint;
+use cairo_lang_starknet_classes::casm_contract_class::{CasmContractClass, CasmContractEntryPoint};
+use cairo_lang_starknet_classes::NestedIntList;
+use cairo_vm::serde::deserialize_program::{
+    ApTracking,
+    FlowTrackingData,
+    HintParams,
+    ReferenceManager,
+};
+use cairo_vm::types::builtin_name::BuiltinName;
+use cairo_vm::types::errors::program_errors::ProgramError;
+use cairo_vm::types::program::Program as CairoVmProgram;
+use cairo_vm::types::relocatable::MaybeRelocatable;
+use serde::de::Error as DeserializationError;
+use serde::{Deserialize, Deserializer};
+use starknet_types_core::felt::Felt;
+
+use crate::core::EntryPointSelector;
+use crate::deprecated_contract_class::{
+    sn_api_to_cairo_vm_program,
+    ContractClass as DeprecatedContractClass,
+    EntryPoint,
+    EntryPointOffset,
+    EntryPointType,
+    Program as DeprecatedProgram,
+};
+
+/// Executable contract class.
+#[derive(Clone, Debug, Eq, PartialEq, derive_more::From)]
 pub enum ContractClass {
+    V0(ContractClassV0),
     V1(ContractClassV1),
 }
 
-/// Compiled contract class variant for Cairo 1 contracts.
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq)]
+pub struct ContractClassV0(pub Arc<ContractClassV0Inner>);
+impl Deref for ContractClassV0 {
+    type Target = ContractClassV0Inner;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq)]
+pub struct ContractClassV0Inner {
+    #[serde(deserialize_with = "deserialize_program")]
+    pub program: CairoVmProgram,
+    pub entry_points_by_type: HashMap<EntryPointType, Vec<EntryPoint>>,
+}
+
+impl TryFrom<DeprecatedContractClass> for ContractClassV0 {
+    type Error = ProgramError;
+
+    fn try_from(class: DeprecatedContractClass) -> Result<Self, Self::Error> {
+        Ok(Self(Arc::new(ContractClassV0Inner {
+            program: sn_api_to_cairo_vm_program(class.program)?,
+            entry_points_by_type: class.entry_points_by_type,
+        })))
+    }
+}
+
+// V1.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ContractClassV1 {
-    Casm(CasmContractClass),
+pub struct ContractClassV1(pub Arc<ContractClassV1Inner>);
+impl Deref for ContractClassV1 {
+    type Target = ContractClassV1Inner;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct ContractClassV1Inner {
+    pub program: CairoVmProgram,
+    pub entry_points_by_type: HashMap<EntryPointType, Vec<EntryPointV1>>,
+    pub hints: HashMap<String, Hint>,
+    bytecode_segment_lengths: NestedIntList,
+}
+
+impl ContractClassV1Inner {
+    #[cfg(any(feature = "testing", test))]
+    pub fn new_for_testing(
+        program: CairoVmProgram,
+        entry_points_by_type: HashMap<EntryPointType, Vec<EntryPointV1>>,
+        hints: HashMap<String, Hint>,
+        bytecode_segment_lengths: NestedIntList,
+    ) -> Self {
+        Self { program, entry_points_by_type, hints, bytecode_segment_lengths }
+    }
+}
+
+impl ContractClassV1 {
+    pub fn bytecode_segment_lengths(&self) -> &NestedIntList {
+        &self.bytecode_segment_lengths
+    }
+}
+
+impl TryFrom<CasmContractClass> for ContractClassV1 {
+    type Error = ProgramError;
+
+    fn try_from(class: CasmContractClass) -> Result<Self, Self::Error> {
+        let data: Vec<MaybeRelocatable> = class
+            .bytecode
+            .into_iter()
+            .map(|x| MaybeRelocatable::from(Felt::from(x.value)))
+            .collect();
+
+        let mut hints: HashMap<usize, Vec<HintParams>> = HashMap::new();
+        for (i, hint_list) in class.hints.iter() {
+            let hint_params: Result<Vec<HintParams>, ProgramError> =
+                hint_list.iter().map(hint_to_hint_params).collect();
+            hints.insert(*i, hint_params?);
+        }
+
+        // Collect a sting to hint map so that the hint processor can fetch the correct [Hint]
+        // for each instruction.
+        let mut string_to_hint: HashMap<String, Hint> = HashMap::new();
+        for (_, hint_list) in class.hints.iter() {
+            for hint in hint_list.iter() {
+                string_to_hint.insert(serde_json::to_string(hint)?, hint.clone());
+            }
+        }
+
+        let builtins = vec![]; // The builtins are initialize later.
+        let main = Some(0);
+        let reference_manager = ReferenceManager { references: Vec::new() };
+        let identifiers = HashMap::new();
+        let error_message_attributes = vec![];
+        let instruction_locations = None;
+
+        let program = CairoVmProgram::new(
+            builtins,
+            data,
+            main,
+            hints,
+            reference_manager,
+            identifiers,
+            error_message_attributes,
+            instruction_locations,
+        )?;
+
+        let mut entry_points_by_type = HashMap::new();
+        entry_points_by_type.insert(
+            EntryPointType::Constructor,
+            convert_entry_points_v1(class.entry_points_by_type.constructor),
+        );
+        entry_points_by_type.insert(
+            EntryPointType::External,
+            convert_entry_points_v1(class.entry_points_by_type.external),
+        );
+        entry_points_by_type.insert(
+            EntryPointType::L1Handler,
+            convert_entry_points_v1(class.entry_points_by_type.l1_handler),
+        );
+
+        let bytecode_segment_lengths = class
+            .bytecode_segment_lengths
+            .unwrap_or_else(|| NestedIntList::Leaf(program.data_len()));
+
+        Ok(Self(Arc::new(ContractClassV1Inner {
+            program,
+            entry_points_by_type,
+            hints: string_to_hint,
+            bytecode_segment_lengths,
+        })))
+    }
+}
+
+#[derive(Clone, Debug, Default, Eq, Hash, PartialEq)]
+pub struct EntryPointV1 {
+    pub selector: EntryPointSelector,
+    pub offset: EntryPointOffset,
+    pub builtins: Vec<BuiltinName>,
+}
+
+impl EntryPointV1 {
+    pub fn pc(&self) -> usize {
+        self.offset.0
+    }
+}
+
+// V0 utilities.
+
+/// Converts the program type from SN API into a Cairo VM-compatible type.
+pub fn deserialize_program<'de, D: Deserializer<'de>>(
+    deserializer: D,
+) -> Result<CairoVmProgram, D::Error> {
+    let deprecated_program = DeprecatedProgram::deserialize(deserializer)?;
+    sn_api_to_cairo_vm_program(deprecated_program)
+        .map_err(|err| DeserializationError::custom(err.to_string()))
+}
+
+// V1 utilities.
+
+// TODO(spapini): Share with cairo-lang-runner.
+/// Used in TryFrom
+fn hint_to_hint_params(hint: &cairo_lang_casm::hints::Hint) -> Result<HintParams, ProgramError> {
+    Ok(HintParams {
+        code: serde_json::to_string(hint)?,
+        accessible_scopes: vec![],
+        flow_tracking_data: FlowTrackingData {
+            ap_tracking: ApTracking::new(),
+            reference_ids: HashMap::new(),
+        },
+    })
+}
+
+/// Used in TryFrom
+fn convert_entry_points_v1(external: Vec<CasmContractEntryPoint>) -> Vec<EntryPointV1> {
+    external
+        .into_iter()
+        .map(|ep| EntryPointV1 {
+            selector: EntryPointSelector(Felt::from(ep.selector)),
+            offset: EntryPointOffset(ep.offset),
+            builtins: ep
+                .builtins
+                .into_iter()
+                .map(|builtin| BuiltinName::from_str(&builtin).expect("Unrecognized builtin."))
+                .collect(),
+        })
+        .collect()
 }
 
 /// All relevant information about a declared contract class, including the compiled contract class
 /// and other parameters derived from the original declare transaction required for billing.
 #[derive(Clone, Debug, Eq, PartialEq)]
+// TODO(Ayelet,10/02/2024): Change to bytes.
 pub struct ClassInfo {
     pub contract_class: ContractClass,
     pub sierra_program_length: usize,
